Circos plot for the Transcription Factors/genes targets.

Generate an plot for Transcription Targets (TFs) or any gene targets. This plot consists of sorting all the chromosomes of any specie based in GTF annotation file and showing how the selected TF(s)/gene(s) targets are distributed. If a target is connected to the same chromosome as the selected one so the connection is defined as cis, otherwise it is a trans connection.

CircosTargets(object, file, nomenclature, selection, cond)

Arguments

object
CeTF class object resulted from runAnalysis function.

file
GTF file or path.

nomenclature
Gene nomenclature: SYMBOL or ENSEMBL.

selection
Specify a single or up to 4 TF/gene to be visualized for.

cond
The options are condition1 or condition2 based on the conditions previously defined in runAnalysis function.

Value

Returns an plot with a specific(s) TF/gene and its targets in order to visualize the chromosome location of each one.

Details

The black links are between different chromosomes while the red links are between the same chromosome.

Examples

if (FALSE) {
CircosTargets(object = out,
file = '/path/to/gtf/specie.gtf',
nomenclature = 'SYMBOL',
selection = 'TCF4',
cond = 'condition1')
}