Partial Correlation and Information Theory (PCIT) analysis

The PCIT algorithm is used for reconstruction of gene co-expression networks (GCN) that combines the concept partial correlation coefficient with information theory to identify significant gene to gene associations defining edges in the reconstruction of GCN.

PCIT(input, tolType = "mean")

Arguments

input
A correlation matrix.

tolType
Type of tolerance (default: 'mean') given the 3 pairwise correlations (see tolerance).

Value

Returns an list with the significant correlations, raw adjacency matrix and significant adjacency matrix.

References

REVERTER, Antonio; CHAN, Eva KF. Combining partial correlation and an information theory approach to the reversed engineering of gene co-expression networks. Bioinformatics, v. 24, n. 21, p. 2491-2497, 2008. https://academic.oup.com/bioinformatics/article/24/21/2491/192682

Examples

# loading a simulated normalized data
data('simNorm')

# getting the PCIT results for first 30 genes
results <- PCIT(simNorm[1:30, ])

# printing PCIT output first 15 rows
head(results$tab, 15)

#>        gene1     gene2    corr1    corr2
#> 1  gene10495 gene10853  0.45665  0.00000
#> 2  gene10495 gene10954 -0.26362  0.00000
#> 3  gene10495   gene113 -0.00966  0.00000
#> 4  gene10495 gene11335  0.48821  0.00000
#> 5  gene10495 gene11474 -0.45507  0.00000
#> 6  gene10495 gene11529  0.24149  0.00000
#> 7  gene10495   gene117 -0.70822 -0.70822
#> 8  gene10495 gene11986  0.23068  0.00000
#> 9  gene10495 gene12969  0.32270  0.00000
#> 10 gene10495 gene13446 -0.04105  0.00000
#> 11 gene10495 gene13600 -0.32514  0.00000
#> 12 gene10495 gene13775  0.44644  0.44644
#> 13 gene10495 gene13933  0.13616  0.00000
#> 14 gene10495 gene14118 -0.52313 -0.52313
#> 15 gene10495 gene14739  0.28042  0.00000