Regulatory Impact Factors (RIF) analysis

The RIF algorithm identify critical transcript factors (TF) from gene expression data.

RIF(input, nta = NULL, ntf = NULL, nSamples1 = NULL, nSamples2 = NULL)

Arguments

input
A matrix of expression with differentially expressed genes and transcript factors in rows, and the samples in columns.

nta
Number of Differentially Expressed (DE) genes.

ntf
Number of Transcription Factors (TFs).

nSamples1
Number of samples of condition 1.

nSamples2
Number of samples of condition 2.

Value

Returns an dataframe with the regulatory impact factors metric for each transcript factor.

Details

The input matrix must have the rows and columns ordered by the following request:

rows: DE genes followed by TFs;

columns: samples of condition1 followed by samples of condition2.

References

REVERTER, Antonio et al. Regulatory impact factors: unraveling the transcriptional regulation of complex traits from expression data. Bioinformatics, v. 26, n. 7, p. 896-904, 2010. https://academic.oup.com/bioinformatics/article/26/7/896/212064

Examples

# load RIF input example
data('RIF_input')

# performing RIF analysis
RIF_out <- RIF(input = RIF_input,
               nta = 104,
               ntf = 50,
               nSamples1 = 10,
               nSamples2 = 10)