Calculate the clustering coefficient

Calculate the clustering coefficient for an adjacency matrix.

clustCoef(mat)

Arguments

mat
An adjacency matrix. Calculating the clustering coefficient only makes sense if some connections are zero i.e. no connection.

Value

Returns the clustering coefficient(s) for the adjacency matrix.

References

Nathan S. Watson-Haigh, Haja N. Kadarmideen, and Antonio Reverter (2010). PCIT: an R package for weighted gene co-expression networks based on partial correlation and information theory approaches. Bioinformatics. 26(3) 411-413. https://academic.oup.com/bioinformatics/article/26/3/411/215002

Examples

# loading a simulated counts data
data('simNorm')

# running PCIT analysis
results <- PCIT(simNorm)

# getting the clustering coefficient
clustCoef(results$adj_sig)

#>  [1]  0.051290806  0.299274803  0.095549335  0.212820701  0.013544699
#>  [6]  0.184294965  0.057548415  0.140254330  0.162951248  0.059506202
#> [11]  0.037321351  0.101274657  0.600455289  0.304748801  0.109650064
#> [16]  0.239657027  0.039740751  0.143921831  0.065823836 -0.001218385
#> [21]  0.108805592  0.054517880  0.063671581  0.119757902  0.096003327
#> [26]  0.201750670  0.132973497  0.076665818  0.049805028  0.227468212
#> [31]  0.115659240  0.107539167  0.111030009  0.050166365  0.198927464
#> [36]  0.039897130  0.093509922  0.012968993  0.096335368  0.073706310
#> [41]  0.224832492  0.147286238  0.055733126  0.016856404  0.369415038
#> [46]  0.091165275  0.041241176  0.080846167  0.068027857  0.087559116
#> [51]  0.025612441  0.162243366  0.037231292  0.075292715  0.045755236
#> [56]  0.125634107  0.230379649  0.040568298  0.217407681 -0.001824641
#> [61]  0.182611804  0.115493566 -0.009074040  0.010672356  0.029068346
#> [66]  0.071195925  0.030178862  0.072913646  0.018137375