Functional Profile of a gene set at specific GO level

Functional Profile of a gene set at specific GO level. Given a vector of genes, this function will return the GO profile at a specific level.

getGroupGO(genes, ont = "BP", keyType, annoPkg, level = 3)

Arguments

genes
Character vector with the genes to perform the functional profile.

ont
One of 'MF', 'BP', and 'CC' subontologies (default: 'BP').

keyType
Key type of inputted genes (i.e. 'ENSEMBL', 'SYMBOL', 'ENTREZID').

annoPkg
Package of annotation of specific organism (i.e. org.Hs.eg.db, org.Bt.eg.db, org.Rn.eg.db, etc).

level
Specific GO Level (default: 3).

Value

Returns an list with the results of the functional profile of the genes and a network with the ontologies (column 1) and the corresponding genes (column 2).

Examples

if (FALSE) { 
# load the annotation package
library(org.Hs.eg.db)

# load the CeTF class object resulted from runAnalysis function
data(CeTFdemo)

# getting the genes in network of condition 1
genes <- unique(c(as.character(NetworkData(CeTFdemo, 'network1')[, 'gene1']),
                 as.character(NetworkData(CeTFdemo, 'network1')[, 'gene2'])))

# performing getGroupGO analysis
cond1 <- getGroupGO(genes = genes,
                    ont = 'BP',
                    keyType = 'ENSEMBL',
                    annoPkg = org.Hs.eg.db, 
                    level = 3)
}