Whole analysis of Regulatory Impact Factors (RIF) and Partial Correlation and Information Theory analysis (PCIT)

This function uses RIF and PCIT algorithms to run the whole pipeline analysis. The pipeline is composed by 4 steps:

Step 1: Data adjustment;

Step 2: Differential expression analysis;

Step 3: Regulatory Impact Factors analysis;

Step 4: Partial Correlation and Information Theory analysis.

runAnalysis(
  mat,
  conditions = NULL,
  lfc = 2.57,
  padj = 0.05,
  TFs = NULL,
  nSamples1 = NULL,
  nSamples2 = NULL,
  tolType = "mean",
  diffMethod = "Reverter",
  data.type = NULL
)

Arguments

mat
Count data where the rows are genes and coluns the samples (conditions).

conditions
A vector of characters identifying the names of conditions (i.e. c('normal', 'tumor')).

lfc
logFoldChange module threshold to define a gene as differentially expressed (default: 2.57).

padj
Significance value to define a gene as differentially expressed (default: 0.05).

TFs
A vector of character with all transcripts factors of specific organism.

nSamples1
Number of samples that correspond to first condition.

nSamples2
Number of samples that correspond to second condition.

tolType
Tolerance calculation type (see tolerance) (default: 'mean').

diffMethod
Method to calculate Differentially Expressed (DE) genes (see expDiff) (default: 'Reverter')

data.type
Type of input data. If is expression (FPKM, TPM, etc) or counts.

Value

Returns an CeTF class object with output variables of each step of analysis.

See also

Examples

data('simCounts')
out <- runAnalysis(mat = simCounts,
                   conditions=c('cond1', 'cond2'),
                   lfc = 3,
                   padj = 0.05,
                   TFs = paste0('TF_', 1:1000),
                   nSamples1 = 10,
                   nSamples2= 10,
                   tolType = 'mean',
                   diffMethod = 'Reverter',
                   data.type = 'counts')